+++ src/libs/overlap.rs
@@ -72,12 +72,20 @@ impl Overlap {
             len: parts[2].parse::<i32>().unwrap(),
             idt: parts[3].parse::<f32>().unwrap(),
 
-            f_strand: parts[4].parse::<i32>().unwrap(),
+            f_strand: if parts[4].parse::<i32>().unwrap() == 0 {
+                0
+            } else {
+                1
+            },
             f_begin: parts[5].parse::<i32>().unwrap(),
             f_end: parts[6].parse::<i32>().unwrap(),
             f_len: parts[7].parse::<i32>().unwrap(),
 
-            g_strand: parts[8].parse::<i32>().unwrap(),
+            g_strand: if parts[8].parse::<i32>().unwrap() == 0 {
+                0
+            } else {
+                1
+            },
             g_begin: parts[9].parse::<i32>().unwrap(),
             g_end: parts[10].parse::<i32>().unwrap(),
             g_len: parts[11].parse::<i32>().unwrap(),
@@ -86,28 +94,54 @@ impl Overlap {
         }
     }
 
-    //    pub fn from_paf(line: &String) -> Self {
-    //        let parts: Vec<&str> = line.split('\t').collect();
-    //
-    //        Self {
-    //            f_id: parts[0].to_string(),
-    //            g_id: parts[5].to_string(),
-    //            len: parts[10].parse::<i32>().unwrap(),
-    //            idt: parts[9].parse::<f32>().unwrap() / parts[10].parse::<f32>().unwrap(),
-    //
-    //            f_strand: 0,
-    //            f_begin: parts[2].parse::<i32>().unwrap() + 1,
-    //            f_end: parts[3].parse::<i32>().unwrap() + 1,
-    //            f_len: parts[1].parse::<i32>().unwrap(),
-    //
-    //            g_strand: parts[8].parse::<i32>().unwrap(),
-    //            g_begin: parts[9].parse::<i32>().unwrap(),
-    //            g_end: parts[10].parse::<i32>().unwrap(),
-    //            g_len: parts[6].parse::<i32>().unwrap(),
-    //
-    //            contained: "overlap".to_string(),
-    //        }
-    //    }
+    pub fn from_paf(line: &str) -> Self {
+        let parts: Vec<&str> = line.split('\t').collect();
+
+        let mut this = Self {
+            f_id: parts[0].to_string(),
+            g_id: parts[5].to_string(),
+            len: parts[10].parse::<i32>().unwrap(),
+            idt: parts[9].parse::<f32>().unwrap() / parts[10].parse::<f32>().unwrap(),
+
+            f_strand: 0,
+            f_begin: parts[2].parse::<i32>().unwrap() + 1,
+            f_end: parts[3].parse::<i32>().unwrap() + 1,
+            f_len: parts[1].parse::<i32>().unwrap(),
+
+            g_strand: 0,
+            g_begin: 0,
+            g_end: 0,
+            g_len: parts[6].parse::<i32>().unwrap(),
+
+            contained: "overlap".to_string(),
+        };
+
+        if parts[4] == "+" {
+            this.g_strand = 0;
+            this.g_begin = parts[7].parse::<i32>().unwrap() + 1;
+            this.g_end = parts[8].parse::<i32>().unwrap() + 1;
+        } else {
+            this.g_strand = 1;
+            this.g_begin = parts[8].parse::<i32>().unwrap() + 1;
+            this.g_end = parts[7].parse::<i32>().unwrap() + 1;
+        }
+
+        if this.f_begin == 1 {
+            this.f_begin = 0;
+        }
+        if this.f_end == 1 {
+            this.f_end = 0;
+        }
+
+        if this.g_begin == 1 {
+            this.g_begin = 0;
+        }
+        if this.g_end == 1 {
+            this.g_end = 0;
+        }
+
+        this
+    }
 
     pub fn to_string(&self) -> String {
         format!(
@@ -175,6 +209,46 @@ fn ovlp_line() {
         assert_eq!(ovlp.g_id(), expected.1);
         assert_eq!(*ovlp.len(), expected.2);
         assert_eq!(*ovlp.idt(), expected.3);
+        assert_eq!(*ovlp.f_strand(), expected.4);
+        assert_eq!(*ovlp.f_begin(), expected.5);
+        assert_eq!(*ovlp.f_end(), expected.6);
+        assert_eq!(*ovlp.f_len(), expected.7);
+        assert_eq!(*ovlp.g_strand(), expected.8);
+        assert_eq!(*ovlp.g_begin(), expected.9);
+        assert_eq!(*ovlp.g_end(), expected.10);
+        assert_eq!(*ovlp.g_len(), expected.11);
+        assert_eq!(ovlp.contained(), expected.12);
+    }
+}
+
+#[test]
+fn paf_line() {
+    let tests = vec![(
+                         "long/5059/0_25030\t25030\t6400\t14738\t+\tlong/9413/0_8928\t8928\t188\t8927\t1427\t8739\t255\tcm:i:168",
+                         (
+                             "long/5059/0_25030",
+                             "long/9413/0_8928",
+                             8739,
+                             0.163,
+                             0,
+                             6401,
+                             14739,
+                             25030,
+                             0,
+                             189,
+                             8928,
+                             8928,
+                             "overlap"
+                         ),
+                     ),
+    ];
+    for (line, expected) in tests {
+        let ovlp = Overlap::from_paf(line);
+        //        assert_eq!(ovlp.to_string(), line);
+        assert_eq!(ovlp.f_id(), expected.0);
+        assert_eq!(ovlp.g_id(), expected.1);
+        assert_eq!(*ovlp.len(), expected.2);
+        assert!((*ovlp.idt() - expected.3).abs() < 0.001);
         assert_eq!(*ovlp.f_strand(), expected.4);
         assert_eq!(*ovlp.f_begin(), expected.5);
         assert_eq!(*ovlp.f_end(), expected.6);